Rename mapping component to mappings
